Workaround for Android 5.x bug in vector icon path data.
In the path data if there is a moveto command with multiple sets of
coordinates Android 5.x misinterprets this and renders the MuPDF
icon incorrectly. SVG states that all coordinate pairs after the first
should be interpreted as implicit lineto commands. When making the
implicit lineto command explicit, Android 5.x is able to draw the
icon correctly.