Update targetSdkVersion and use built-in file chooser to open files.
This dances around with an awkward LibraryActivity to fire up the
file chooser. There's probably a better way to do this directly with
the DocumentActivity, but I'll figure that out another day.

Also fix the insane mess that is the intent-filter in AndroidManifest.xml.
Register for all the mime-types we handle (some of which android
recognizes) and also for application/octet-stream so we can catch files
that Android doesn't (yet) recognize.