Do not start file chooser activity unnecessarily.
When returning to the LibraryActivity it is important to know whether
the file chooser activity has already been started. Previously this
information was not saved, so if the LibraryActivity was resumed then
a new file chooser activity would be started. This caused a user on
Android 13 and later to have to click the back button more than once
to exit the LibraryActivity.

This commit adds a boolean indicating whether the file chooser activity
has already been started. If not already started, then a file chooser
activity is started, otherwise the LibraryActivity just continues to
wait for a result from the file chooser activity.