Use insets to avoid system bar and page seek bar overlap.
Inspired by:
https://developer.android.com/develop/ui/views/layout/edge-to-edge#system-bars-insets

We start listening to WindowInsets, and when they arrive we record
and union the display cut-out and system bar insets. This way we can
avoid selfie cameras interfering with the search bar at the top of the
screen and system bars from interfering with the page scrubber at the
bottom of the screen.

We introduce a topBar and rename the navigationBar the bottomBar. The
topBar contains the actionBar and searchBar, while the bottomBar
contains the page scrubber. For the insets to take effect we setup the
topBar to listen to the top insets while the bottomBar listens to the
bottom insets.

Moreover we move the background color from the actionBar and searchBar
to the topBar. This way the entire topBar will have the same
background color, both in the inset area and in the action button
area.

When the bottomBar is enabled and showing the first page of a document
it is difficult to grab hold of the thumb when it is all the way to
the left at the page scrubber without also triggering a Back gesture.

In order solve this page scrubber vs. system gestures interference we
must set a system gesture exclusion rectangle. These are capped at
200dp vertical height, but there doesn't appear to be any horizontal
restriction. We set an exclusion rectangle that covers the entire
bottom system bar of the page scrubber and equally far up the into the
page view area. The reason for the latter is that it is easy for a
large finger to accidentally start in the page view area and slide
down onto the page scrubber to change page, and you don't want this
slight mistake to start a back gesture either.

This has all been tested with Android versions back to minSdkVersion,
i.e. Android 5.0, API level 21.